tg-client-bot-gpt/requirements.txt
absl-py==2.2.2
annotated-types==0.7.0
anyio==4.9.0
astunparse==1.6.3
asyncpg==0.30.0
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
distro==1.9.0
et_xmlfile==2.0.0
filelock==3.18.0
flatbuffers==25.2.10
fsspec==2025.3.2
gast==0.6.0
google-pasta==0.2.0
grpcio==1.71.0
h11==0.14.0
h5py==3.13.0
hf-xet==1.0.3
httpcore==1.0.8
httpx==0.28.1
httpx-socks==0.10.0
huggingface-hub==0.30.2
idna==3.10
jiter==0.9.0
keras==3.9.2
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
namex==0.0.8
numpy==2.1.3
openai==1.75.0
openai-proxy==0.2.5
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.15.0
packaging==24.2
pandas==2.2.3
protobuf==5.29.4
pyaes==1.6.1
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
Pyrogram==2.0.106
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-socks==2.7.1
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
safetensors==0.5.3
six==1.17.0
sniffio==1.3.1
termcolor==3.0.1
TgCrypto==1.2.5
tiktoken==0.9.0
tokenizers==0.21.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
wrapt==1.17.2
XlsxWriter==3.2.3
